fix: Client의 weight, height를 Double->Float로 변경
Former-commit-id: 7d883e6f3c7c64d9c512a803d0845f7fcdc5045a

+++ BE/src/main/java/com/S209/yobi/domain/measures/helper/tbl_bodycomp.java
@@ -10,8 +10,8 @@ import lombok.Getter;
 public class tbl_bodycomp {
     private int uAge;
     private String uGender; // 예: "male" 또는 "female"
-    private double uHeight;
-    private double uWeight;
+    private float uHeight;
+    private float uWeight;
     private double bfp;       // 체지방률
     private double bfm;
     private double bmr;       // 기초대사량
